+++ cmdline/standalone_cli_tools/restore_files/RestoreFilesCmdMain.cpp
@@ -40,7 +40,7 @@ int main(const int argc, char *const *const argv) {
   }
   cta::log::StdoutLogger log(hostName, "cta-restore-deleted-files");
 
-  cta::admin::RestoreFilesCmd cmd(std::cin, std::cout, std::cerr, log);
+  cta::cliTool::RestoreFilesCmd cmd(std::cin, std::cout, std::cerr, log);
   int ret = cmd.main(argc, argv);
   // Delete all global objects allocated by libprotobuf
   google::protobuf::ShutdownProtobufLibrary();